Create CartmanSpam
This will set the PC's volume to max, then open a ton of browser tabs which will play a YouTube video of Eric Cartman singing "She Works Hard for the Money." I UNDERSTAND there are TONS of easier ways to do this. This is one of my earliest DuckyScripts and I've purposely tried to use the KISS method, so that anyone, including me, can understand what's happening here. If you want more spam, just copy and paste the repeated section a few more times. Enjoy!